增加了方法和参数支持：在GUI中新增method和eta参数配置，支持二次采样方法；更新了采样逻辑，修复了time_steps的类型问题，优化了参数规范性和展示布局。

+++ scheduler/linear_noise_scheduler.py
@@ -230,7 +230,7 @@ class DDIMSampler(nn.Module):
             a = self.T // steps
             time_steps = np.asarray(list(range(0, self.T, a)))
         elif method == "quadratic":
-            time_steps = (np.linspace(0, np.sqrt(self.T * 0.8), steps) ** 2).astype(np.int)
+            time_steps = (np.linspace(0, np.sqrt(self.T * 0.8), steps) ** 2).astype(np.int32)
         else:
             raise NotImplementedError(f"sampling method {method} is not implemented!")
